Reorganize the whole repository and add yaruMate
