+++ lib/yaru.dart
@@ -1,5 +1,8 @@
 library yaru;
 
-export 'package:yaru/src/theme.dart';
-export 'package:yaru/src/utils/colors.dart';
-export 'package:yaru/src/utils/flavor_colors.dart';
+export 'package:yaru/src/colors/yaru_colors.dart';
+export 'package:yaru/src/colors/flavor_colors.dart';
+export 'package:yaru/src/themes/yaru_light.dart';
+export 'package:yaru/src/themes/yaru_dark.dart';
+export 'package:yaru/src/themes/yaru_mate_light.dart';
+export 'package:yaru/src/themes/yaru_mate_dark.dart';
